+++ Public/OSDCloudTS/Test-DISMFromOSDCloudUSB.ps1
@@ -8,6 +8,17 @@
     )
     $OSDCloudUSB = Get-Volume.usb | Where-Object {($_.FileSystemLabel -match 'OSDCloud') -or ($_.FileSystemLabel -match 'BHIMAGE')} | Select-Object -First 1
     if ($OSDCloudUSB){
+        $OSDCloudDriveLetter = $OSDCloudUSB.DriveLetter
+    }
+    $MappedDrives = (Get-CimInstance -ClassName Win32_MappedLogicalDisk).DeviceID | Select-Object -Unique
+    if ($MappedDrives){
+        ForEach ($MappedDrive in $MappedDrives){
+            if (Test-Path -Path "$MappedDrive\OSDCloud"){
+                $OSDCloudDriveLetter = $MappedDrive.replace(":","")
+            }
+        }
+    }
+    if ($OSDCloudDriveLetter){
         $ComputerProduct = (Get-MyComputerProduct)
         if (!($PackageID)){
             $PackageID = $DriverPack.PackageID
@@ -15,8 +26,8 @@
         }
         $ComputerManufacturer = (Get-MyComputerManufacturer -Brief)
         if ($ComputerManufacturer -match "Samsung"){$ComputerManufacturer = "Samsung"}
-        $DriverPathProduct = "$($OSDCloudUSB.DriveLetter):\OSDCloud\DriverPacks\DISM\$ComputerManufacturer\$ComputerProduct"
-        $DriverPathPackageID = "$($OSDCloudUSB.DriveLetter):\OSDCloud\DriverPacks\DISM\$ComputerManufacturer\$PackageID"
+        $DriverPathProduct = "$($OSDCloudDriveLetter):\OSDCloud\DriverPacks\DISM\$ComputerManufacturer\$ComputerProduct"
+        $DriverPathPackageID = "$($OSDCloudDriveLetter):\OSDCloud\DriverPacks\DISM\$ComputerManufacturer\$PackageID"
         if ($PackageID){  
             Write-Host "Testing Paths:"
             Write-Host "  $DriverPathProduct"