+++ RELEASENOTES.md
@@ -10,10 +10,17 @@ Note, the changes I make won't go into effect until the next module release date
 
 ## Changes
 
+
+
 ### 24.3.12 (not yet implemented)
 - Start-OSDCloudGUIDev Updates
   - Dynamically pulling the OS Edition and Index Numbers based on the OS Name, Language and Activation chosen in the GUI.
     - This is to improve experience on other language esd files [Issue 117](https://github.com/OSDeploy/OSD/issues/117)
+- Updated DISMFromOSDCloudUSB function to now search mapped network drives too
+  - If you have a network drive that follows the OSDCloud USB Drive folder structure, it will use that as a cache similar to the flash drive
+    - F:\OSDCloud\DriverPacks\DISM\HP\859C
+    - F:\OSDCloud\OS\
+    - etc
   
 ### 24.3.6 (implemented in OSD Module 24.3.10.1)
 - Promoted code from Start-OSDCloudGUIDev to Start-OSDCloudGUI